make Member page a dynamic route
